# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Products_Store",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/prize": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves all prizes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query string",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/prize.ResponsePrize"
                            }
                        }
                    },
                    "400": {
                        "description": "There is no matches with query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create new prize",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "json_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RequestPrize"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/prize.ResponsePrize"
                        }
                    },
                    "400": {
                        "description": "Field validation failed on the 'required' tag",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/promo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves all promo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query string",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/promo.ResponsePromo"
                            }
                        }
                    },
                    "400": {
                        "description": "There is no matches with query",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create new promo",
                "parameters": [
                    {
                        "description": "json body",
                        "name": "json_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.RequestPromo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/promo.ResponsePromo"
                        }
                    },
                    "400": {
                        "description": "There is no category with id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/promo/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Retrieves promo by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/promo.ResponsePromo"
                        }
                    },
                    "400": {
                        "description": "There is no product with id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete promo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product with id was deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "There is no product with id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.RequestPrize": {
            "type": "object",
            "required": [
                "description"
            ],
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "main.RequestPromo": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "participant.Participant": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "prize.Prize": {
            "type": "object"
        },
        "prize.ResponsePrize": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "promo.ResponsePromo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/participant.Participant"
                    }
                },
                "prizes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/prize.Prize"
                    }
                }
            }
        }
    }
}